+++ server/src/utils/ZetaSqlUtils.ts
@@ -3,6 +3,7 @@ import { AnyASTTypeProto__Output } from '@fivetrandevelopers/zetasql/lib/types/z
 import { ParseLocationRangeProto__Output } from '@fivetrandevelopers/zetasql/lib/types/zetasql/ParseLocationRangeProto';
 import { StructFieldProto } from '@fivetrandevelopers/zetasql/lib/types/zetasql/StructFieldProto';
 import { TypeProto } from '@fivetrandevelopers/zetasql/lib/types/zetasql/TypeProto';
+import { ColumnDefinition } from '../TableDefinition';
 
 export function positionInRange(position: number, range: ParseLocationRangeProto__Output): boolean {
   return range.start <= position && position <= range.end;
@@ -14,6 +15,42 @@ export function rangeContainsRange(outerRange: ParseLocationRangeProto__Output, 
 
 export function rangesEqual(range1: ParseLocationRangeProto__Output, range2: ParseLocationRangeProto__Output): boolean {
   return range1.start === range2.start && range1.end === range2.end;
+}
+
+// TODO: add unit tests
+export function createType(newColumn: ColumnDefinition): TypeProto {
+  const bigQueryType = newColumn.type.toLowerCase();
+  const typeKind = TypeFactory.SIMPLE_TYPE_KIND_NAMES.get(bigQueryType);
+  let resultType: TypeProto;
+  if (typeKind) {
+    resultType = {
+      typeKind,
+    };
+  } else if (bigQueryType === 'record') {
+    resultType = {
+      typeKind: TypeKind.TYPE_STRUCT,
+      structType: {
+        field: newColumn.fields?.map<StructFieldProto>(f => ({
+          fieldName: f.name,
+          fieldType: createType(f),
+        })),
+      },
+    };
+  } else {
+    console.log(`Cannot find TypeKind for ${newColumn.type}`); // TODO: fix all these issues
+    resultType = {
+      typeKind: TypeKind.TYPE_STRING,
+    };
+  }
+  if (newColumn.mode?.toLocaleLowerCase() === 'repeated') {
+    resultType = {
+      typeKind: TypeKind.TYPE_ARRAY,
+      arrayType: {
+        elementType: resultType,
+      },
+    };
+  }
+  return resultType;
 }
 
 export function toTypeProto(astType: AnyASTTypeProto__Output): TypeProto | undefined {
